Use Perl From PATH
Instead of the hard-coded perl, use the perl that comes first in PATH.
This particularly important when more than one perl is installed (e.g.
a system one, and a user one). The assumption is that the perl which
comes first in PATH will be set up properly (@INC and stuff). Mixing
perl versions leads to very obscure error messages. It is hence
paramount to use the same perl executable, that was used to install
the modules.